refactoring CatalogRepository.cs and ArticleRepository.cs files

+++ ANTWebAPI/src/ANTWebAPI.DAL/Repositories/ArticleRepository.cs
@@ -13,6 +13,7 @@ public class ArticleRepository(ANTDbContext context) : IRepository<Article>
     {
         if (await _context.Contents.AnyAsync(e => e.ArticleId == id)) return false;
         var model = await _context.Articles.FirstAsync(e => e.Id == id);
+        if (model == null) return false;
         _context.Articles.Remove(model);
         await _context.SaveChangesAsync();
         return true;
